Refactor shared persistence contracts for store extraction
